Move test classes out of package to root directory
